qcacld-3.0: Fix the issue in processing the del_tspec req
SME qos expects the RIC data and RIC data length to process the
add_ts response after roaming and updates the qos state to QOS_ON
after the successful processing. RIC data is included in re-assoc
response frame and updated in pe session.

Since the RIC data info is not given to SME qos in roam sync event
processing, qos state is not updated, hence the del_tspec command
is buffered and not processed. Send the RIC data info to SME qos
in roam sync processing so that the SME qos state is updated properly
and processes any buffered commands.

Change-Id: I7551b0bab4f11fe2d68c014cf89834ba199ad5ff
CRs-Fixed: 1010374