qcacld-3.0: Convert wlan_hdd_hostapd.c to unified logging [3/7]
Currently the HDD code uses a variety of logging APIs.  In qcacld-3.0
HDD should converge on a unified set of logging APIs.  Update
wlan_hdd_hostapd.c to use the unified set of APIs.  Since there are a
large number of changes in this file, perform the conversion in
phases.

For phase 3 update the following functions:
__iw_softap_setparam()
__iw_softap_get_three()

Change-Id: I5db86f54bcc978d7935873423fa1e91bc83d7a2e
CRs-Fixed: 937661