qcacld-3.0: aquire mutex qdf_conc_list_lock only if required
The mutex qdf_conc_list_lock should be used to protect the
access of conc_connection_list. If driver holds this lock and calls
sme api, then there is a chance of entering into a deadlock scenario
as sme api is going to aquire the sme global lock.

This change is going to make sure that driver doesn't hold
qdf_conc_list_lock and call any sme api.

CRs-Fixed: 1086929
Change-Id: I93819d13905609a6a2886b77e951db4253f281bd