qcacld-3.0: Enable DFS channel for a single channel scan
In csr_scan_copy_request_valid_channels_only(), the DFS channels
are ignored if ini parameter gSkipDfsChannelInP2pSearch is enabled.
If the supplicant issues a single channel scan for a DFS channel,
driver will do a full scan. As its a single channel scan which is a
DFS channel would be ignored and the num of channels to scan would
be zero.

And if the number of channels to scan are zero, driver
does a full scan which would impact the overall connection
time.

To address this, enable DFS channel for a single
channel scan.

Change-Id: I166b9674a6cb6bb9c180c1c8d68480440ea867be
CRs-Fixed: 2038974