qcacmn: Fix cfg_in_range for non-ini values
Currently, the macro type and the "is an ini" value information are
clubbed to together in the cfg component. This lead to not being able to
use the macro type with non-ini configuration values. Split macro type
and "is an ini" value information in the cfg component so that
cfg_in_range can be used with non-ini configuration values.

Change-Id: If63f222cf21a45f0d7d3809f1be845ecc2b2cd65
CRs-Fixed: 2278656