qcacld-3.0: De-init adapters before stop modules
In con_mode handler, adapters are getting deinitialized
while cleaning up present mode which is happening after
stop modules. In stop modules driver checks for memory
leak which results in mem leak detection for the adapters
which are not freed yet.

To address this issue, move the deinitialization of the
adapters from clean up present mode to stop present mode, which
happens before stop modules.

Change-Id: I9cda456feb2b0c199f5c8d4debee21f0e07c47fb
CRs-Fixed: 2326340