qcacld-3.0: Reject SAP channel change req only in connecting state
Channel change req for SAP is rejected even when STA is waiting
for key. In case of force SCC, if STA is connecting to security
profile, During SAP channel change STA is waiting for key and
thus force SCC is rejected.

Fix is to reject the SAP channel change only when STA is in
connecting state.

Change-Id: I7ff8ed8487ec8d996f9fb4b1ffe2ac303ba235df
CRs-Fixed: 2388842