qcacld-3.0: Replace typedef tSirTsmIE
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSirTsmIE typedef does not meet
any of those criteria, so replace it (and the "tp" variant) with a
reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names
and so-called Hungarian notation, so in conjunction rename the
underlying struct to be in compliance.

Change-Id: I9c0784f93fa67b571bb816ded4a137f4435accfc
CRs-Fixed: 2394249