qcacld-3.0: Update wiphy band info after FW ready event is received
The driver can be configured to support only 2.4 GHz operation
via INI setting "BandCapability=1". Currently, when the driver
is configured in this manner, the RF bands attribute in the
WPS IE in Probe Request frames incorrectly indicates the driver
can support both 2.4 GHz and 5 GHz operation. This is because
the logic used to configure the wiphy->bands occurs before the
INI file has been processed, and hence the BandCapability is
not properly applied. To address this issue relocate the logic
that updates the wiphy->bands until after the INI file has been
processed.

Wiphy supported band info is also dependent on FW capabilities,
so update wiphy band info after FW ready event is received.

Change-Id: Ic210cafd730f56feb9df06d8ad5fcd8cca8355fc
CRs-Fixed: 2377343