qcacld-3.0: Remove check of enable 2x2 in hdd_update_nss
Currently the driver checks enable 2x2 in hdd update nss, and
if it is 0, i.e nss is max limited to 1, then the driver returns
from there with the fail, thus not changing the global ini.

Fix is to remove the check from the common path, and put the
check only in the path where the FW also supports per vdev nss
change command.

Change-Id: I164e052eeec05d8e8c1653d134f645689f9e6281
CRs-Fixed: 2409355