qcacld-3.0: Fix Mlme info updating in connected BSS in scan
Mlme info is not updated for the new AP to which driver roamed and
also not updated for the old AP after driver roamed to new AP.
Also in some case when deauth is from peer the Mlme info is not reset
by driver for the disconnected AP.
As mlme info of the AP is not updated and it remain in associated state
The AGE out logic skip the AP and thus it never ageout.

Fix this by properly updating the MLME info of the AP during roaming
And disconnection.

Change-Id: Ib17499c8ab0c4725d146d7c09077c92661adbdd5
CRs-Fixed: 2419853