qcacld-3.0: Modify PCL based on SRD ini for P2P-GO
Currently the driver does not update the PCL
channel list based upon the SRD ini, and thus
adds the SRD channels in the PCL for p2p-go,
which can lead to p2p-go operation on a channel
not expected.

Fix is to remove the SRD channels from the PCL
channel list, if the SRD ini is set.

Change-Id: I8824fa38ef59976816a5e230982c551263da0c9c
CRs-Fixed: 2437795