qcacld-3.0: Fix P2P-GO restart fail issue in unsafe channel event
Currently the driver checks MCC/SCC with the present connections,
and band interference in restart case of both SAP, and P2P-GO,
which can lead to p2p-go fail, as supplicant which is the
initiator of the switch channel does not care about MCC, and wants
the driver to restart the p2p-go in the channel given.

Fix is to overide the checks for P2P-GO only, and restart the
vdev in the given channel by supplicant.

Change-Id: I5db1da5452272696d80128e16b9e0ee50aa0de28
CRs-Fixed: 2466292