qcacmn: Enable radar detection and subchannel marking in Agile DFS
When radar is found on Agile detector, the channels are not added to
NOL list and are only added to preCAC NOL list.

Add the radar affected channels, in Agile detector, to NOL list and to
preCAC NOL list based on the detector ID of the radar event received.

For subchannel marking, use agile channel instead of current operating
channel to the NOL when radar is found on agile detector.

Change-Id: Ifa61feeddbaaa81fe405972aa5826994a1383c00
CRs-Fixed: 2464929