qcacld-3.0: Set correct op_flag in sme_set_del_pmkid_cache
Currently the host driver sets PMKSA op_flag for flush and
delete PMKSA entry as in both the cases PMK length is 0.
This is not correct because for delete PMKSA command only
the bssid and SSID will be sent to the firmware.
For the host driver to distinguish between del_pmksa and
flush pmksa, the is_flush_all flag should be used.
is_flush_all: true - Flush pmksa
is_flush_all: false - set/del pmksa based on the action flag.

Change-Id: I7eac65317a9c6904cfd08acce328a65df451269e
CRs-Fixed: 2503372