qcacld-3.0: Prevent wlan suspend if conn in progress
Currently the driver does not check if the state
is key exchange in progress and suspends wlan
before set-key happens which further results in
FW also in wake up state.
This would result in delayed EAP exchange, and also
in power loss.

Scenario:-
1. Turn on STA and try to connect to enterprise network
2. Turn off display.

Here the display turn off would trigger APPS suspend
while the STA is trying to connect, and authenticate
and since there is no check in driver to prevent
suspend in set key in progress state, it would result
in a FW assert, as the expectation of FW is to allow
suspend only after set key has been done.

Fix is to prevent WLAN-suspend in case of connection
in progress, and allow suspend only in connected
and authenticated state.

Change-Id: Ic173116f7ba424005d938a43c75831a6a4dc874c
CRs-Fixed: 2512866