qcacld-3.0: Adjust the intervel of tsf sync timer
Because it has clock drift between Qtimer and target timer, so host
driver need to get two timer offset periodically, then transform the
tsf time from each RX packet to qtime according to this offset value.
If the intervel is too large, the result will not very accurate.
Modify the value from 9 seconds to 3 seconds for STA mode.
Modify the SAP tsf sync intervel to be the same as STA mode.

Change-Id: Id915cbf386059df8233246e196279cb376e2a25b
CRs-Fixed: 2531940