ASoC: wcd-mbhc: Fix headset insert not detected after moisture config
Headset insertion is not detected after moisture is configured and then
removed. This is because during moisture_detect GND_DET_EN is configured
irrespective of whether it is enabled or not. Add a check before enabling
gnd_det_en. Also as moisture_en is disabled and duty cycle is enabled
change check during removal to disable threshold and polling.

Change-Id: Iee9d79b38b228e78d1509d59e92d37f1328f079e
Signed-off-by: Vatsal Bucha <vbucha@codeaurora.org>