qca-wifi: CFR: Update channel capture status counters
PHY may not be able to capture the CFR data in certain cases.
Ucode will update the reason due to which CFR capture did not
happen in PHYRX_PKT_END TLV.

Channel capture status -
	0 - CAPTURE_IDLE
	1 - CAPTURE_BUSY
	2 - CAPTURE_ACTIVE
	3 - CAPTURE_NO_BUFFER

Originally, ucode updated channel capture status only for PPDUs,
for which PHY was able to capture CFR data successfully. Ucode is
making a change to report channel capture status in PPDUs for which
PHY was unable to capture CFR data.

Re-positioning the counters in host to handle this new change in
ucode.

Change-Id: I915fe7c4969649754936cab624343e044b397a0c