qcacld-3.0: Do not fill 6GHz channels for legacy command
Currently when driver receives the legacy get channel command,
it fills all the channels which includes 6GHz channels also.
This may lead to unexpected behaviour for the applications which
does not have support for the 6GHz channels.

To address above issue, do not add 6Ghz channels in the output of
the get channels command.

Change-Id: I87beea4e80906846fa5d1d1076792385d4424297
CRs-Fixed: 2782638