qcacmn: Expand the scope of the check "radar on Agile detector"
The "radar on Agile detector" currently checks only if the detector
id in the radar found parameter is the same as the Agile detector id.

Increase the scope of the "radar on Agile detector" check by also
checking if the Agile PreCAC was enabled or if RCAC was enabled and
if PreCAC timer was running.

CRs-Fixed: 2770006
Change-Id: I064560f738bdcbbf4bcb45b37bae71a0829e4146