qcacld-3.0: Fix lim deferred messages not getting processed
When channel Switch is happening and north bound disconnect is
also initiated at same time before vdev restart response is
received, the disconnect command will be activated by serialization
module but the eWNI_SME_DEAUTH_REQ message to lim was deferred.
The deferred message queue at lim was not processed as there was
no wmi command response received.

So process the deferred messages after the gLimProcessDefdMsgs
is set to 1.

Change-Id: I8200407663857349ea2eef0924fe1c345fab8028
CRs-Fixed: 2789632