qcacld-3.0: Free the memory for radio stats in case of timeout
Currently host driver does not free the memory for radio stats
if there is a timeout which may lead to unexpected behavior.

For example if there are multiple radio stats events from fw
and if it receives only one event and does not receives next
events in that case host allocates the memory for the channels
received in the first event and does not free this memory in wma
as it expects remaining events, since remaining events will not
come host gets timeout in hdd and memory in wma is not freed.
when host issues next command for the ll_stats it receives the
response for this command and as memory is already allocated for
channel stats for the previous event it combines this new response
with the old response which results in unexpected response to the
userspace.

To address above issue, add mem free logic to free the memory for
channel stats at wma.

Change-Id: I6735f8c282c4205c11548aabe3c8e269e81e4e54
CRs-Fixed: 2813478