disp: msm: sde: prepare encoders during enable commit
Prepare commit hook is getting skipped for encoders during
power-on. This is required for clearing the pending flush
mask but the encoder->crtc link is not established at
this point for an enable commit.

Instead of looking at the old CRTC state and matching an
attached encoder, use the encoder_mask in the new CRTC state
to determine which encoders to prepare.

Change-Id: I0754020abd328504899d9591180dedc1723aae49
Signed-off-by: Steve Cohen <cohens@codeaurora.org>