qcacmn: Update the max_bw for every reg channel
Update the maximum bandwidth value of the channels in the current
channel list based on the chip's supported modes and the
availability of the adjacent channels.

Also remember the maximum phymode and bandwidth supported by
the device in the regulatory pdev private object.

For determining if a channel is DFS or not, use the flag
REGULATORY_CHAN_RADAR which comes from regulatory rules, instead of
using CHANNEL_STATE_DFS.

Change-Id: I844857ef63af3f7628b37eedfe5098bb88e3e0dc