qca-wifi: Use regulatory channel list to build precac list
The current method to build the precac list uses the ic channels, an
exhaustive list of all possible channel phymodes and channel widths to
calculate the maximum possible channel width of a given channel number.
Since this inefficient use the regulatory channel list to get the maximum
possible channel width and build the precac list.

Change-Id: I89544cf1b3d60b73038edd1df02243857782c514
CRs-Fixed: 2832062