qcacld-3.0: Don't use proto_generic_nss for SAP
Taking minimun of proto generic nss and dynamic
nss is only valid for STA as proto generic nss
is minimum of self capability and peer AP
capability. So, while fetching nss for SAP,
it is returning as 0.

So, in case of SAP don't take minimum of proto
generic nss with dynamic nss in case of SAP.

Change-Id: I82fabd346da11087d6c6a2123faa52f0a55adf35
CRs-Fixed: 2836723