msm: ipa: Fix doing polling after NAPI reschedule
Make changes to do polling in same NAPI context after
NAPI is rescheduled as we can be stuck forever in
same NAPI context.

Make changes to not assert in case clock is not active
when handling TX interrupt as the packet would have been
processed already as part of NAPI reschedule.

Change-Id: I65c87fa31899a745ff99cc7708eed79927263815
Signed-off-by: Chaitanya Pratapa <cpratapa@codeaurora.org>