qcacld-3.0: Use makefile flag to select DLKM makefile
The kernel is transitioning from AndroidKernelModule.mk to
Build_external_kernelmodule.mk as the helper makefile for building
DLKMs. As the new makefile was introduced for the "taro" platform,
initially a platform check was introduced to determine which makefile
to include.  However this is not a scalable approach.

Recently the flag TARGET_USES_KERNEL_PLATFORM was added to the build
system, so update the logic such that when this flag is set to 'true'
we will use Build_external_kernelmodule.mk.

Change-Id: I2200b27753263b931cebdbff5445d50b1fc2b94f
CRs-Fixed: 2913254