qcacmn: Add support to send channel info to WIN LOWI application
In the current code, host does not send the channel list information to
WIN LOWI application. LOWI application populates the TX power in OEM RTT
request frame from the host received channel list information. Since the
channel list was not received from host TX power was set to 0. This was
leading to low RSSI for RTT ack frames.

Send the channel list information of all the PDEVs and PSOCs to WIN LOWI
application  to set a valid TX power.

Change-Id: I9b7d740b37349d092629aa2f9550555901fb4331