qcacld-3.0: Don't allow set_fast_roaming if roaming is in progress
When roaming/roam sync is in progress, the host receives a "disable
fast roaming" indication from the supplicant. Host process roaming
or roam sync indication in schedular thread and process disable
fast roaming process is in a supplicant thread. As both processes
are executed in different threads, there is a possibility in the
host that the roaming state gets changes from ROAM_SYNCH_IN_PROG
to RSO_STOPPED or ROAMING_IN_PROG to RSO_STOPPED in the middle of
roaming.

A state change in the middle of roaming can lead to multiple
issues.

To avoid this host should not process disable roaming indication
from userspace, if roaming is in progress.

Change-Id: I9b02683eab7db9b5be12ada6d922073e5267b8a5
CRs-Fixed: 2944303