qcacld-3.0: Change roam state after connect complete
With commit I84c77644c6023322ac77e307b06be70ec0f58ed5, Host
allows dual sta roaming only if two STA's sessions for are
active (STA + STA) and concurrency is in DBS.

Currently, before the increment of the number of active
sessions in the policy manager, the host tries to enable
dual-band roaming. This results in dual sta roaming failure
with the reason number of active sessions equal to one.

The fix is to make sure the host should try to enable dual-band
roaming after incrementing the number of active sessions in the
policy manager.

Change-Id: I150926897cb773db979279edc54e5fb731d98e43
CRs-Fixed: 2971831