qcacld-3.0: Use static memory allocation for DFS
In WLAN ON/OFF stress test, due to system memory fragment, it's hard
to dynamiclly allocate contiguous memory for DFS huge memory which
is greater than 64K. Normally memory allocations of order of 4 and
upwards are not recommended by memory team. So use static memory
allocation for DFS.

ROME and Hastings already changed, this change is for Geona.

Change-Id: I33aa3efd378518491f06b496344ed3034c4978d9
CRs-Fixed: 2851914