qcacld-3.0: Fix getting wrong nss issue for dual APs case
In the dual APs DBS case, hdd_get_sap_operating_band
always return the operating band of last adapter and
might not be the current adapter's operating band,
which is not right.
So adding this fix to get operating frequency of
current adapter and then convert to according band,
then get the correct nss info.

Change-Id: Ief171d9b3d66937bede5fbe65be031bdc19529c8
CRs-Fix: 3043735