qcacld-3.0: Add ultra high bandwidth voting for dbs mode
Add a new SNOC/DDR voting slot for DBS mode. In DBS KPI tests,
more hardware resources are needed for both HW MAC0/1.
This change adds a new logic to vote more DDR frequency when
it's in DBS mode and the overall throughput is larger than a
pre-defined threshold.

Change-Id: I14a318b417a39714865de0617c9943c1a5e04ae6
CRs-fixed: 3064197