qcacld-3.0: Refactor the SB disconnect allowed logic in LIM
Currently each SB disconnect API, uses similar but not exact
logic to decide if disconnect is already is progress and
can the new disconnect be allowed.

Move this code to a new API lim_is_sb_disconnect_allowed and
use the same API from all the places so that logic remain exact
the same.

Also add the check at few missing places.

Change-Id: If3c966ff0f36cb7dffe99ef33bfe1ec37e7a6fb2
CRs-Fixed: 3076049