qcacmn: Handle unsupported EAPOL packet type
In function qdf_fill_wlan_connectivity_log, the unsupported EAPOL
packet types are not dropped.This results in setting the
log_subtype to invalid value zero and the log buffer being forwarded
to wlan_connectivity_log_enqueue. The fix is to drop the unsupported
packets by returning from the function.

The local variable log_buf,of type struct wlan_log_record, is not
null initialized.This could cause some of the  member elements to have
unexpected values.This fix is to null initialize the log_buf structure.

Update Qualcomm copyright to include year 2022

Change-Id: I651329fead1e42282d7811c32eea2359ba970527
CRs-Fixed: 3079947