qcacmn: Fix invalid btm candidate TLV conditional check
The number of BTM candidate info TLV is sanitized against the
total number of candidate info in the roam stats event. But
the condition for validation is not correct. So BTM request
candidates and preferences are logged as 0.

Fix invalid btm candidate TLV conditional check in
extract_roam_trigger_stats_tlv.

Change-Id: Ib96cd8c2ba84726b23490bcaf8a934cca2c82d2c
CRs-Fixed: 3142568