qcacld-3.0: Register latency level event call back handler
Issue : Host fails to register the event call back handler
again in case Wi-Fi is turned on and triggered a reboot.

Because even after reboot, at a point of time when the host
is registering event callback handler, the host does not
receive FW capability flags yet in mlme context and this
results in FW caps check failure every time while call back
registration.

The fix is to allow the host to register to call back handler
irrespective of FW flags and add capability check in callback
handler.

Change-Id: I1dd66279097db6a68885b0416076428b73eefb9c
CRs-Fixed: 3196030