qcacld-3.0: Fix the channel for M2 packet in pkt capture mode
Currently in packet capture, channel for data packets is updated
from pdev. We are updating channel in pdev when we receive auth
frame in mon thread. But, sometimes we are updating channel in
radiotap of M2 in pkt_capture_update_tx_status, before mon thread
process auth packet and update channel in pdev, because of
which we are updating wrong channel in M2.

So, to avoid this move update of channel from mon thread to
pkt_capture_mgmtpkt_process.

Change-Id: Icf14c4331849ad7cfdfbd7a3043e4e4dc3d33721
CRs-Fixed: 3266890