disp: msm: sde: add static display support to neo target
DPU HW is overriding the SW configured cache read attribute to always
read with invalidate, when color format is fsd. New scid was
introduced where the read attribute is always set to normal read at
LLCC, disregarding the attribute from DPU.
This change adds support of new scid 16, for static display in neo
target. This change also disables the general cache state machine
sequence, as it is not required for neo target.

Change-Id: Ib1acc7b4e0ab544e9ae5e2d1f077b52b1a8fdfc0
Signed-off-by: Andhavarapu Karthik <quic_kartkart@quicinc.com>