qcacld-3.0: Change sap Channel if unsafe during cac
While doing CAC for moving SAP to DFS channel, channel
avoid event can be received. While in CAC is going on,
the current sap channel is not updated and remains the
channel on which sap was before CAC. It is only updated
after CSA. So, when channel avoid event is received,
it is checked against the previous frequency and hence,
no action is taken as the previous frequency is part of
safe channel. This results in sap starting and remaining
in Unsafe channel.
As, a fix when sap starts after CSA, check for unsafe
channel and if the sap frequency is unsafe then restart
it to another channel.

Change-Id: I910b80fe87fc149f25e84383b128a5e5c9d269e4
CRs-Fixed: 3287239