qcacld-3.0: Reduce latency in Roam sync processing
Upon the completion of roam sync indication event,
the roam sync complete cmd is first sent to the firmware.
This is followed by the netdev queues enable. This
approach adds latency in the first downstream tx frame,
and the FT roaming KPI is not met due to this latency.

To fix this, if roam sync handling is successful,
enable the queues and then send roam sync complete
to the firmware.

This change improves roaming latency by 6 msec.

Change-Id: Ib05ca01892b6d8b62badcafb64bb8c16270ee9f8
CRs-Fixed: 3483751