qcacmn: Add a entry for hw version of QCA6490 in qwlan_hw_list
Currently, driver does not fill correct target name for
target_version = 0x400C0210 as the entry corresponding to this
target version is not present in qwlan_hw_list[] table.
So, add the entry in qwlan_hw_list[] for above target version
so that driver fills proper target name for it.

Change-Id: I490a6ced3a43e3e08f1502a83dd99ac35d2f251d
CRs-Fixed: 3774962