qcacld-3.0: Fix update of roam band during connection
Currently the roam_band_bitmask is being updated only during vdev
create as part of rso init. However, if the band cap gets changed
from userspace after vdev creation, the roam_band_bitmask is not
being updated. Therefore, after connection, the stale bitmask is
used to configure RSO and PCL lists to the firmware.

To fix this, update the roam band bitmask as part of the user space
operation directly, so that proper bandmask can be used after connection.

Change-Id: I8d14bdb0064ba051b3edc308f7505276db1aa5e3
CRs-Fixed: 3809636